+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 39.88.0 - 2022-01-20
+* [#2415](https://github.com/stripe/stripe-dotnet/pull/2415) API Updates
+  * Add support for new resource `PaymentLink`
+  * Add support for `PaymentLink` on `CheckoutSession`
+
+## 39.87.0 - 2022-01-19
+* [#2414](https://github.com/stripe/stripe-dotnet/pull/2414) API Updates
+  * Add support for `BacsDebit` and `Eps` on `PaymentIntentPaymentMethodOptionsOptions`, `PaymentIntentPaymentMethodOptionsOptions`, `PaymentIntentPaymentMethodOptionsOptions`, and `PaymentIntentPaymentMethodOptions`
+  * Add support for `ImageUrlPng` and `ImageUrlSvg` on `PaymentIntentNextActionWechatPayDisplayQrCode`
+
 ## 39.86.0 - 2022-01-13
 * [#2413](https://github.com/stripe/stripe-dotnet/pull/2413) API Updates
   * Add support for `PaidOutOfBand` on `Invoice`